Split index-editor components to separate files and small fixes
